// src/utils/enums/Sorting.ts
enum Sorting {
  TitleAscending = "Title (ascending)",
  TitleDescending = "Title (descending)",
  DateAscending = "Date (ascending)",
  DateDescending = "Date (descending)",
  CreatedDateAscending = "Date created (ascending)",
  CreatedDateDescending = "Date created (descending)",
}

export default Sorting;